requirements.txt
# Web framework
flask>=3.1.0
gunicorn>=23.0.0

# logging
google-cloud-logging>=3.11.3

# Core dependencies
sentence-transformers>=3.3.1
numpy>=2.1.3
pandas>=2.2.3
scikit-learn>=1.5.2

# Firebase dependencies
firebase-admin>=6.6.0
google-cloud-firestore>=2.19.0

# Environment and testing
python-dotenv>=1.0.1
matplotlib>=3.9.2

# Google Cloud dependencies
google-cloud-logging

# GRPC dependencies
grpcio>=1.68.0
